Cleanup

+++ src/com/bekvon/bukkit/residence/utils/ParticleEffects.java
@@ -1146,7 +1146,7 @@ public enum ParticleEffects {
 		 */
 		@Override
 		public float getValueX() {
-			return (float) red / 255F;
+			return red / 255F;
 		}
 
 		/**
@@ -1156,7 +1156,7 @@ public enum ParticleEffects {
 		 */
 		@Override
 		public float getValueY() {
-			return (float) green / 255F;
+			return green / 255F;
 		}
 
 		/**
@@ -1166,7 +1166,7 @@ public enum ParticleEffects {
 		 */
 		@Override
 		public float getValueZ() {
-			return (float) blue / 255F;
+			return blue / 255F;
 		}
 	}
 
@@ -1204,7 +1204,7 @@ public enum ParticleEffects {
 		 */
 		@Override
 		public float getValueX() {
-			return (float) note / 24F;
+			return note / 24F;
 		}
 
 		/**
